Making setup script work for arm, too

+++ setup.sh
@@ -1,17 +1,31 @@
+#!/bin/bash
+
 echo -e "\nWhat's your email address (used for Let's Encrypt SSL cert registration)?"
 
 read email
 
-echo -e "\n-- installing wget and sudo if needed\n"
-apt-get update
-apt-get install -y wget sudo
+if ! which wget
+  then echo -e "\n-- installing wget\n" ; sudo apt-get update ; sudo apt-get install -y wget
+fi
+
+arch=`uname -m`
+
+case $arch in
+  x86_64)
+    suffix="x64"
+    ;;
+  armv6l)
+    suffix="armv6"
+    ;;
+esac
 
 cd ~
-echo -e "\n-- downloading altcloud executable\n"
-wget https://altcloud.io/linux-x64/altcloud
-chmod a+x altcloud
-sudo mv altcloud /usr/local/bin
-
+if [ ! -f /usr/local/bin/altcloud ]; then
+  echo -e "\n-- downloading altcloud executable for $arch\n"
+  wget https://altcloud.io/linux-$suffix/altcloud
+  chmod a+x altcloud
+  sudo mv altcloud /usr/local/bin
+fi
 
 # set up altcloud user
 echo -e "\n-- setting up altcloud user\n"
@@ -20,9 +34,10 @@ sudo setcap cap_net_bind_service=+ep `readlink -f \`which altcloud\``
 
 echo -e "\n-- setting up altcloud service\n"
 # set up service
-sudo cat > /etc/systemd/system/altcloud.service <<- EOM
+sudo tee /etc/systemd/system/altcloud.service <<- EOM
 [Service]
-ExecStart=/usr/local/bin/altcloud server /home/altcloud/webroot
+ExecStart=/usr/local/bin/altcloud server
+WorkingDirectory=/home/altcloud/webroot
 Restart=always
 StandardOutput=syslog
 StandardError=syslog
@@ -40,8 +55,8 @@ echo -e "\n-- creating webroot, keys, config, index page (as altcloud)\n"
 sudo -u altcloud mkdir -p /home/altcloud/webroot
 cd /home/altcloud/webroot
 sudo -u altcloud altcloud keys
-sudo -u altcloud echo -e "Hello world!" > index.html
-sudo -u altcloud cat > .config <<- EOM
+echo "Hello world!" | sudo -u altcloud tee index.html
+sudo -u altcloud tee .config <<- EOM
 letsencrypt:
   email: $email
 EOM
@@ -49,7 +64,7 @@ EOM
 # switch back to root, start service
 cd ~
 echo -e "\n-- starting altcloud service\n"
-systemctl enable altcloud
-systemctl start altcloud
+sudo systemctl enable altcloud
+sudo systemctl start altcloud
 
 echo -e "done!"
